+++ packages/inspector/src/components/Assets/MobileSession/utils.ts
@@ -1,0 +1,5 @@
+export function fmt(n: number, decimals = 1): string {
+  if (n >= 1_000_000) return (n / 1_000_000).toFixed(decimals) + 'M';
+  if (n >= 1_000) return (n / 1_000).toFixed(decimals) + 'K';
+  return decimals > 0 ? n.toFixed(decimals) : String(n);
+}
